refactor: Merge types/docs from `basic.pyi` into `crudclient/auth/basic.py`
- Merged type hints and docstrings from `crudclient/auth/basic.pyi` into `crudclient/auth/basic.py`.

- Deleted `crudclient/auth/basic.pyi`.
